Basic business detail page and menu page in place. Usable enough to provide users with a lot of data! Woohoo!

+++ menu/views.py
@@ -51,19 +51,42 @@ def bus_signup(request):
     return HttpResponseRedirect('/menu/'+str(b.id))
   else:
     form = InterestedBusinessForm()
-  return render_to_response('menu/merc_signup.html',{'form':form}, context_instance=RequestContext(request))    
+  return render_to_response('menu/merchant/signup.html',{'form':form}, context_instance=RequestContext(request))    
+
 
 @login_required
-def fill_menu(request,bid):
+def get_menulist(request):
+  businesses = InterestedBusiness.objects.all().select_related('menuitem')
+  return render_to_response('menu/menus.html',{'businesses':businesses}, context_instance=RequestContext(request))    
+ 
+
+@login_required
+def get_edit_business(request,bid):
+  b = get_object_or_404(InterestedBusiness,pk=int(bid))
+  editing = True
+  businessform = InterestedBusinessForm(instance=b)
+  if request.method == 'POST':
+    if request.user != b.owner:
+      return HttpResponseRedirect('/invalid')
+    bform = InterestedBusinessForm(request.POST, instance=b)
+    if bform.is_valid():
+      bform.save()
+      return HttpResponseRedirect('/menu/'+str(b.id))
+
+  return render_to_response('menu/menuelements/menu.html',{'business':b,'busform': businessform, 'editing':editing}, context_instance=RequestContext(request))    
+    
+@login_required
+def get_edit_menu(request,bid):
    form = MenuItemForm()
    allergyform =AllergyInfoForm()
    otherinfoform = OtherInfoForm()
    nutritionform = NutritionInfoForm()
    b = get_object_or_404(InterestedBusiness,pk=int(bid))
    if request.method == 'POST':
+     if request.user != b.owner:
+       return HttpResponseRedirect('/invalid')
      mi = MenuItem(business=b)
      form = MenuItemForm(request.POST,instance=mi)
-     print form
      if form.is_valid():
        menuitem = form.save()
        instanceAllergy = AllergyInfo(menuitem=menuitem)
@@ -79,9 +102,9 @@ def fill_menu(request,bid):
        print 'saved a form!'
      else:
        print 'invalid form'
-    
+     return HttpResponseRedirect('/menu/'+str(b.id))
    mealName = generate_menu_list(b)
-   return render_to_response('menu/menu.html',{'business':b,'menuitems':mealName, 'mform':form, 'nform':nutritionform, 'aform':allergyform, 'oiform': otherinfoform}, context_instance=RequestContext(request))    
+   return render_to_response('menu/menuelements/menu.html',{'business':b,'menuitems':mealName, 'mform':form, 'nform':nutritionform, 'aform':allergyform, 'oiform': otherinfoform}, context_instance=RequestContext(request))    
 
 @login_required
 def edit_menu(request,bid,mid):
@@ -97,6 +120,8 @@ def edit_menu(request,bid,mid):
   nutritionform = NutritionInfoForm( instance=instanceNutrition)
   print ' before and method is ' + str(request.method)
   if request.method == 'POST':
+    if request.user != b.owner:
+      return HttpResponseRedirect('/invalid')
     form = MenuItemForm(request.POST,instance=mi)
     print 'here and form is ' + str(form)
     if form.is_valid():
@@ -116,19 +141,58 @@ def edit_menu(request,bid,mid):
       print 'invalid form'
     return HttpResponseRedirect('/menu/'+str(b.id))
   mealName = generate_menu_list(b,int(mid))
-  return render_to_response('menu/menu.html',{'business':b,'menuitems':mealName, 'mform':menuitemform, 'nform':nutritionform, 'aform':allergyform, 'oiform': otherinfoform, 'editing':True}, context_instance=RequestContext(request))    
+  return render_to_response('menu/menuelements/menu.html',{'business':b,'menuitems':mealName, 'mform':menuitemform, 'nform':nutritionform, 'aform':allergyform, 'oiform': otherinfoform, 'editing':True}, context_instance=RequestContext(request))    
 
 @login_required
 def delete_menu(request,bid,mid):
-   menuitemform = MenuItemForm()
-   allergyform =AllergyInfoForm()
-   otherinfoform = OtherInfoForm()
-   nutritionform = NutritionInfoForm()
-   MenuItem.objects.filter(id=mid).delete()
-   
    b = get_object_or_404(InterestedBusiness,pk=int(bid))
-   
-   mealName = generate_menu_list(b)
-   return render_to_response('menu/menu.html',{'business':b,'menuitems':mealName, 'mform':menuitemform, 'nform':nutritionform, 'aform':allergyform, 'oiform': otherinfoform}, context_instance=RequestContext(request))    
+   if request.user != b.owner:
+     return HttpResponseRedirect('/invalid')
+   MenuItem.objects.filter(id=mid).delete()
+   return HttpResponseRedirect('/menu/'+str(b.id))
 
+
+@login_required
+def get_edit_details(request,bid,mid):
+  b = get_object_or_404(InterestedBusiness,pk=int(bid))
+  mi = MenuItem.objects.get(id=mid)
+  instanceAllergy = mi.allergy
+  instanceNutrition = mi.nutrition
+  instanceOther = mi.otherinfo
+    
+
+  if 'editing' in request.GET:
+    if request.user != b.owner:
+     return HttpResponseRedirect('/invalid')
+    editing = True
+  else:
+    editing = False
+
+  form = MenuItemForm(instance=mi)
+  allergyform = AllergyInfoForm( instance=instanceAllergy)
+  otherinfoform = OtherInfoForm(  instance=instanceOther)
+  nutritionform = NutritionInfoForm( instance=instanceNutrition)
+  print ' before and method is ' + str(request.method)
+  if request.method == 'POST':
+    if request.user != b.owner:
+     return HttpResponseRedirect('/invalid')
+    form = MenuItemForm(request.POST,instance=mi)
+    print 'here and form is ' + str(form)
+    if form.is_valid():
+      mi = form.save()
+      instanceAllergy = mi.allergy
+      instanceNutrition = mi.nutrition
+      instanceOther = mi.otherinfo
+      allergyform = AllergyInfoForm(request.POST, instance=instanceAllergy)
+      otherinfoform = OtherInfoForm(request.POST, instance=instanceOther)
+      nutritionform = NutritionInfoForm(request.POST, instance=instanceNutrition)
+      print nutritionform
+      allergyform.save()
+      nutritionform.save()
+      otherinfoform.save()
+      print 'saved a form!'
+    else:
+      print 'invalid form'
+    return HttpResponseRedirect('/menu/'+str(b.id)+'/details/'+str(mi.id))
+  return render_to_response('menu/dish/dishdetails.html',{'business':b,'m':mi, 'mform':form, 'nform':nutritionform, 'aform':allergyform, 'oiform': otherinfoform, 'editing':editing}, context_instance=RequestContext(request))
